Fix sandboxed preload module format
