+++ apps/desktop/electron/main.ts
@@ -356,7 +356,7 @@ async function createWindow() {
     titleBarStyle: "hiddenInset",
     backgroundColor: "#f8f6f1",
     webPreferences: {
-      preload: path.join(__dirname, "preload.js"),
+      preload: path.join(__dirname, "preload.cjs"),
       contextIsolation: true,
       sandbox: true
     }
